+++ NEWS.md
@@ -9,9 +9,11 @@ border-radius: 128px;
 ## v1.8.0
 
 - Features
-  - Support MetronInfo.xml v1.0 Schema for comic metadata. Only metadata
-    features that already exist in Codex.
+  - Support MetronInfo.xml v1.0 Schema for comic metadata but only metadata
+    features that already exist in Codex for now.
   - Support CB7 (7zip) comic archives.
+  - Skip updating comics if the file updated but it looks like the metadata did
+    not.
   - Uses Comicbox 2.0 see
     [Comicbox NEWS for all details](https://github.com/ajslater/comicbox/blob/main/NEWS.md)